@redbaron 2016-09-18
I  don't really get why there are 1024 mutexes minimum and 4096 mutexes by default, but maybe these numbers can be lowered now?