@redbaron 2016-09-18
or possibly embed mutexes into metric (or wrapped metric struct) to be more cache friendly. From the look of it every time mutex is accessed in local/storage.go there is a matching access to a metric itself, so keeping then adjastent is natural then.

You'd also wont need hash + modulo anymore as every metric would have its own lock, making locking more fain grained