@juliusv 2016-09-18
Embedding the locks in the metrics wouldn't work, as there can be multiple `model.Metric`s for the same time series flying around, and we also need to be able to lock a series according to its metric fingerprint before the series exists. Also, then we'd have millions of locks instead of just a couple - not a huge deal, but more memory wasted.

No idea about the 1024 / 4096 minimum / maximum. @beorn7 can probably say more about that.

Introducing the padding is great though.