@beorn7 2016-09-18
> No idea about the 1024 / 4096 minimum / maximum.

The 1024 is just a safety guard so that nobody accidentally sets a lower number and gets sever performance penalties. The rationale is that a lower number would save so little RAM that it doesn't make sense to go down there.

The 4096 default is there because it is still pretty cheap in terms of RAM but it helps a lot for high frequency ingestion with a decent number of time series. (With more time series, you need to tweak this value much higher.)

With this change, the RAM consumption increases by 8x. Still I think we should not go below 4096 by default as this is what "normal" users will need anyway.