@redbaron 2016-09-19
@beorn7 my concern wasn't RAM usage, but L1 cache size which is just 32KB