@beorn7 2016-09-19
Feel free to play with lower numbers. My rationale so far is that you need high performance if you have high frequency injection and/or high number of series, resulting in a lot of concurrency. Reducing the number of FP mutexes will create more contention, which will probably kill the caching effects of using fewer locks. I could be wrong, of course.